Split context propagation package
