+++ internal/contextpropagation/context_propagator.go
@@ -4,18 +4,11 @@ import (
 	"context"
 
 	"github.com/cschleiden/go-workflows/backend/metadata"
+	"github.com/cschleiden/go-workflows/contextpropagation"
 	"github.com/cschleiden/go-workflows/internal/sync"
 )
 
-type ContextPropagator interface {
-	Inject(context.Context, *metadata.WorkflowMetadata) error
-	Extract(context.Context, *metadata.WorkflowMetadata) (context.Context, error)
-
-	InjectFromWorkflow(sync.Context, *metadata.WorkflowMetadata) error
-	ExtractToWorkflow(sync.Context, *metadata.WorkflowMetadata) (sync.Context, error)
-}
-
-func Inject(ctx context.Context, metadata *metadata.WorkflowMetadata, propagators []ContextPropagator) error {
+func Inject(ctx context.Context, metadata *metadata.WorkflowMetadata, propagators []contextpropagation.ContextPropagator) error {
 	for _, propagator := range propagators {
 		err := propagator.Inject(ctx, metadata)
 		if err != nil {
@@ -26,7 +19,7 @@ func Inject(ctx context.Context, metadata *metadata.WorkflowMetadata, propagator
 	return nil
 }
 
-func Extract(ctx context.Context, metadata *metadata.WorkflowMetadata, propagators []ContextPropagator) (context.Context, error) {
+func Extract(ctx context.Context, metadata *metadata.WorkflowMetadata, propagators []contextpropagation.ContextPropagator) (context.Context, error) {
 	for _, propagator := range propagators {
 		var err error
 		ctx, err = propagator.Extract(ctx, metadata)
@@ -38,7 +31,7 @@ func Extract(ctx context.Context, metadata *metadata.WorkflowMetadata, propagato
 	return ctx, nil
 }
 
-func InjectFromWorkflow(ctx sync.Context, metadata *metadata.WorkflowMetadata, propagators []ContextPropagator) error {
+func InjectFromWorkflow(ctx sync.Context, metadata *metadata.WorkflowMetadata, propagators []contextpropagation.ContextPropagator) error {
 	for _, propagator := range propagators {
 		err := propagator.InjectFromWorkflow(ctx, metadata)
 		if err != nil {
@@ -49,7 +42,7 @@ func InjectFromWorkflow(ctx sync.Context, metadata *metadata.WorkflowMetadata, p
 	return nil
 }
 
-func ExtractToWorkflow(ctx sync.Context, metadata *metadata.WorkflowMetadata, propagators []ContextPropagator) (sync.Context, error) {
+func ExtractToWorkflow(ctx sync.Context, metadata *metadata.WorkflowMetadata, propagators []contextpropagation.ContextPropagator) (sync.Context, error) {
 	for _, propagator := range propagators {
 		var err error
 		ctx, err = propagator.ExtractToWorkflow(ctx, metadata)
@@ -65,12 +58,12 @@ type propagatorsKey int
 
 var propagatorsCtxKey propagatorsKey
 
-func WithPropagators(ctx sync.Context, propagators []ContextPropagator) sync.Context {
+func WithPropagators(ctx sync.Context, propagators []contextpropagation.ContextPropagator) sync.Context {
 	return sync.WithValue(ctx, propagatorsCtxKey, propagators)
 }
 
-func Propagators(ctx sync.Context) []ContextPropagator {
-	propagators, ok := ctx.Value(propagatorsCtxKey).([]ContextPropagator)
+func Propagators(ctx sync.Context) []contextpropagation.ContextPropagator {
+	propagators, ok := ctx.Value(propagatorsCtxKey).([]contextpropagation.ContextPropagator)
 	if !ok {
 		return nil
 	}
