+++ workflow/activity.go
@@ -6,7 +6,7 @@ import (
 	"github.com/cschleiden/go-workflows/backend/metadata"
 	a "github.com/cschleiden/go-workflows/internal/args"
 	"github.com/cschleiden/go-workflows/internal/command"
-	"github.com/cschleiden/go-workflows/internal/contextpropagation"
+	icontextpropagation "github.com/cschleiden/go-workflows/internal/contextpropagation"
 	"github.com/cschleiden/go-workflows/internal/converter"
 	"github.com/cschleiden/go-workflows/internal/fn"
 	"github.com/cschleiden/go-workflows/internal/sync"
@@ -65,9 +65,9 @@ func executeActivity[TResult any](ctx Context, options ActivityOptions, attempt 
 	name := fn.Name(activity)
 
 	// Capture context
-	propagators := contextpropagation.Propagators(ctx)
+	propagators := icontextpropagation.Propagators(ctx)
 	metadata := &metadata.WorkflowMetadata{}
-	if err := contextpropagation.InjectFromWorkflow(ctx, metadata, propagators); err != nil {
+	if err := icontextpropagation.InjectFromWorkflow(ctx, metadata, propagators); err != nil {
 		f.Set(*new(TResult), fmt.Errorf("injecting workflow context: %w", err))
 		return f
 	}
